Add weighted rankings, PPR, sortable leaderboards, data import fixes
- Replace PPD with PPR (Points Per Round = PPD x 3) across all views
- Add sortable column headers to Player and Team Leaderboard tables
- Add weighted ranking: dense rank with Out/In avg tiebreakers, weights (AS*2 + Out*2 + In + Close) / 6
- Add OutAvg/InAvg to leaderboard API for ranking tiebreakers
- Fix Cricket MPR: MarksScored INT -> DECIMAL(6,2)
- Fix All-Stars: store allStarCount in JSON Details, SUM instead of COUNT
- Add historical data import script (scripts/import-history.js)
- Add season champion banner, player theme colors, settings page
- Add live match page, announcer utility
- Exclude History/ folder from git

+++ database/001_create_schema.sql
@@ -169,7 +169,7 @@ CREATE TABLE Turns (
     IsGameOut       BIT             NOT NULL DEFAULT 0,     -- closing turn of X01
 
     -- Cricket / Shanghai fields
-    MarksScored     INT             NULL,                   -- total marks this turn
+    MarksScored     DECIMAL(6,2)    NULL,                   -- total marks this turn
     IsCricketClose  BIT             NOT NULL DEFAULT 0,     -- winning dart in Cricket
 
     -- Shanghai extra fields
